refactor: Everything is broken rn. I'm refactoring everything
Added a few files and removed the atlas_runtime. `atlas_runtime` will re-appear once the language is stable, it'll be used to run `comptime` expression like in Zig/Cpp (`constexpr`). I'll update the roadmap a bit. I hope I can have a semi-working version for the end of the week