feat: Start of the runtime rework
I started laying the ground for the rework with my current thoughts. The instruction set is still not well defined. The stack will see a lot of changes, from a plain Vec<VMData> to a Vec<StackFrame> that each holds a little exec_stack as well as an array of variables. The memory (heap) will be changed for the better but not heavily refactored. Structs will make their appearance with the reworked stack