feat: atlas_codegen should be fixed and atlas_asm has been added.
The assembler isn't really working yet, for now it only outputs a `Vec<u8>` of the instructions and nothing else, no object descriptor, constant pool, or whatever, but it's coming soon enough. I'll use serde to help me with and I think the instructions will use u32 or something with and immediate value will be done on the remaining 24 bits of data. If it's bigger than that, then just go and grab it from the constant pool, ez.