fix: Issue with the typechecker & some examples
The compiler still doesn't compile down to the asm::Instruction, and I still need to redo most, if not all, the Codegen system, and also add back in the HIR & Typecheker all the new expressions. This should be done quite easily, so no issue with that. (The VM still doesn't run at all lmao)