fix(self): Fixed an issue in the brainfuck lexer...
I was able to forget how Vector works, silly me. I did
ew Vector<Token>(new [Token; 8]) thinking it would make the capacity 8, but not the length, and welp... It actually makes both 8, which means I was pushing at index 8, 9, ... but not at the start of the array. I am stupid.